requirements.txt
aiohttp==3.11.18
redis==6.1.0
python-dotenv==1.1.0
aiohttp_socks==0.10.1